chore: split brainstorming repository into multipler smaller repositories
